LibJS: Move SafeFunction into the Heap folder
As the heap has special knowledge about SafeFunction - and the entire
point of this function is to teach the GC about SafeFunction's captures.
