+++ Userland/Libraries/LibJS/Heap/SafeFunction.h
