+++ Userland/Libraries/LibJS/Runtime/VM.h
@@ -18,6 +18,7 @@
 #include <LibJS/CyclicModule.h>
 #include <LibJS/Heap/Heap.h>
 #include <LibJS/Heap/MarkedVector.h>
+#include <LibJS/Heap/SafeFunction.h>
 #include <LibJS/ModuleLoading.h>
 #include <LibJS/Runtime/CommonPropertyNames.h>
 #include <LibJS/Runtime/Completion.h>
