+++ Userland/Libraries/LibWeb/Fetch/Infrastructure/FetchAlgorithms.h
@@ -10,7 +10,6 @@
 #include <LibJS/Heap/Cell.h>
 #include <LibJS/Heap/GCPtr.h>
 #include <LibJS/Heap/HeapFunction.h>
-#include <LibJS/SafeFunction.h>
 #include <LibWeb/Forward.h>
 
 namespace Web::Fetch::Infrastructure {
