+++ Userland/Libraries/LibWeb/HTML/SessionHistoryTraversalQueue.h
@@ -13,7 +13,6 @@
 #include <LibJS/Heap/CellAllocator.h>
 #include <LibJS/Heap/GCPtr.h>
 #include <LibJS/Heap/HeapFunction.h>
-#include <LibJS/SafeFunction.h>
 #include <LibWeb/Forward.h>
 
 namespace Web::HTML {
